+++ src_py/agenthub/claude5/client.py
@@ -204,10 +204,14 @@ class Claude5Client(LLMClient):
                     content_blocks.append(await self._convert_image_url_to_source(item["image_url"]))
                 elif item["type"] == "thinking":
                     if item["thinking"] == REDACTED_THINKING:
-                        content_blocks.append({"type": "redacted_thinking", "data": item["signature"]})
+                        content_blocks.append({"type": "redacted_thinking", "data": item["fidelity"]["signature"]})
                     else:
                         content_blocks.append(
-                            {"type": "thinking", "thinking": item["thinking"], "signature": item["signature"]}
+                            {
+                                "type": "thinking",
+                                "thinking": item["thinking"],
+                                "signature": item["fidelity"]["signature"],
+                            }
                         )
                 elif item["type"] == "tool_call":
                     content_blocks.append(
@@ -263,7 +267,9 @@ class Claude5Client(LLMClient):
                     {"type": "partial_tool_call", "name": block.name, "arguments": "", "tool_call_id": block.id}
                 )
             elif block.type == "redacted_thinking":
-                content_items.append({"type": "thinking", "thinking": REDACTED_THINKING, "signature": block.data})
+                content_items.append(
+                    {"type": "thinking", "thinking": REDACTED_THINKING, "fidelity": {"signature": block.data}}
+                )
 
         elif claude_event_type == "content_block_delta":
             event_type = "delta"
@@ -277,7 +283,7 @@ class Claude5Client(LLMClient):
                     {"type": "partial_tool_call", "name": "", "arguments": delta.partial_json, "tool_call_id": ""}
                 )
             elif delta.type == "signature_delta":
-                content_items.append({"type": "thinking", "thinking": "", "signature": delta.signature})
+                content_items.append({"type": "thinking", "thinking": "", "fidelity": {"signature": delta.signature}})
 
         elif claude_event_type == "content_block_stop":
             event_type = "stop"